+++ app/Http/Controllers/KepalaToko/AkunController.php
@@ -62,7 +62,18 @@ class AkunController extends Controller
     {
         $item = User::findOrFail($id);
 
+        if (
+            $item->relasiService()->exists() || $item->relasiSale()->exists()
+            || $item->relasiAssembly()->exists() || $item->incident()->exists() || $item->expense()->exists()
+            || $item->salary()->exists()
+        ) {
+            toast('Data Akun yang memiliki riwayat transaksi tidak bisa dihapus.', 'error');
+            return redirect()->back();
+        }
+
         $item->delete();
+
+        toast('Data Akun berhasil dihapus.', 'success');
 
         return redirect()->route('akun');
     }
